result evaluated with rouge metric
It stands for Recall-Oriented Understudy for Gisting Evaluation, which sounds quite a fancy term at the first look. Simply put, it is a metric to measure the quality of an automatically produced summary or translation against a reference or a set of references (human-produced) summary or translation. There are several variations of ROUGE